+++ CHANGELOG.rst
@@ -29,7 +29,7 @@ Change Log
 - [???] "asynch" multienv
 - [???] properly model interconnecting powerlines
 
-[1.6.4] - 2021-xx-yy
+[1.6.4] - 2021-11-yy
 ---------------------
 - [BREAKING] the name of the python file for the "agent" module are now lowercase (complient with PEP). If you
   did things like `from grid2op.Agent.BaseAgent import BaseAgent` you need to change it like
@@ -38,6 +38,8 @@ Change Log
 - [FIXED] a bug where the shunt had a voltage when disconnected using pandapower backend
 - [FIXED] a bug preventing to print the action space if some "part" of it had no size (empty action space)
 - [FIXED] a bug preventing to copy an action properly (especially for the alarm)
+- [FIXED] a bug that did not "close" the backend of the observation space when the environment was `closed`. This 
+  might be related to `Issue#245 <https://github.com/rte-france/Grid2Op/issues/255>`_
 - [ADDED] serialization of `current_iter` and `max_iter` in the observation.
 - [ADDED] the possibility to use the runner only on certain episode id
   (see `runner.run(..., episode_id=[xxx, yyy, ...])`)
@@ -49,6 +51,8 @@ Change Log
 - [IMPROVED] documentation to clearly state that the action_class should not be modified.
 - [IMPROVED] possibility to tell which chronics to use with the result of `env.chronics_handler.get_id()` (this is also
   compatible in the runner)
+- [IMPROVED] it is no more possible to call "env.reset()" or "env.step()" after an environment has been closed: a clean error
+  is raised in this case.
 
 [1.6.3] - 2021-08-21
 --------------------